[svn-r230] Changes were actually made by Robb.  I am commiting them for him while he is visiting LLNL.  I changed the default creation template offset and length to 4.  Will fix the problem later.
Changes since 19980205
----------------------

./src/H5H.c
./src/H5Hprivate.h
./src/H5O.c
./src/H5Ocont.c
./src/H5Odtype.c
./src/H5Oefl.c
./src/H5Olayout.c
./src/H5Oname.c
./src/H5Onull.c
./src/H5Oprivate.h
./src/H5Odspace.c
./src/H5Ostab.c
./src/debug.c
./html/H5.format.html
        Added an extra 4-byte field after the heap magic number for
        alignment on the DEC alpha. Changed object header message
        alignment to 8-bytes.

./src/H5F.c
./src/H5Farray.c
./src/H5Ffamily.c
./src/H5Fistore.c
./src/H5Flow.c
./src/H5Fprivate.h
./src/H5Fsec2.c
./src/H5Fstdio.c
./src/H5Gnode.c
./src/H5O.c
./src/H5Odtype.c
./src/H5P.c
./src/H5Pprivate.h
./src/H5T.c
./src/H5Tconv.c
./src/H5Tpkg.h
./src/H5Tpublic.h
./src/H5V.c
./src/H5detect.c
./test/cmpd_dset.c
./test/dsets.c
./test/dtypes.c
./test/extend.c
./test/hyperslab.c
./test/istore.c
./test/th5p.c
./test/theap.c
        Fixed a few irix64 warnings regarding size_t vs. int,
        variables set but not used, printf formats

./config/irix64
        Added `-woff 1196' to get rid of errors about __vfork() being
        implicitly defined in a system header file.

./src/H5B.c
        Fixed a stack alignment problem.

+++ src/H5Tpublic.h
@@ -209,7 +209,7 @@ H5T_str_t H5Tget_strpad (hid_t type_id);
 intn H5Tget_nmembers (hid_t type_id);
 char *H5Tget_member_name (hid_t type_id, int membno);
 size_t H5Tget_member_offset (hid_t type_id, int membno);
-int H5Tget_member_dims (hid_t type_id, int membno, int dims[]/*out*/,
+int H5Tget_member_dims (hid_t type_id, int membno, size_t dims[]/*out*/,
                         int perm[]/*out*/);
 hid_t H5Tget_member_type (hid_t type_id, int membno);
 H5T_cset_t H5Tget_cset (hid_t type_id);
